% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_function.R
\name{bayes}
\alias{bayes}
\title{Bayes function when running mcem with mcmc setting}
\usage{
bayes(
  person.data = NA,
  person.id = "",
  task.id = "",
  max.counts = "",
  obs.counts = "",
  time = "",
  parallel = T,
  n.chains = NA,
  thin = 1,
  iter = NA,
  burn = NA
)
}
\arguments{
\item{person.data}{- student reading data}

\item{person.id}{The column name in the data that represents the unique individual identifier.}

\item{task.id}{The column name in the data that represents the unique task identifier.}

\item{max.counts}{The column name in the data that represents the number of words in a task.}

\item{obs.counts}{The column name in the data that represents the words read correctly for each case.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{parallel}{parallel=T, #logical, run in parallel? "T" or "F"}

\item{n.chains}{int., number of the chains}

\item{thin}{int, thinning interval, a.k.a, period of saving samples}

\item{iter}{int., number of the iterations after the burn-in period}

\item{burn}{int., number of the burn-in iterations}
}
\value{
list
}
\description{
Bayes function when running mcem with mcmc setting
}
