% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.wcpm_function.R
\name{bayes.wcpm}
\alias{bayes.wcpm}
\title{Bayes function when running mcem with mcmc setting}
\usage{
bayes.wcpm(
  calib.data = NA,
  stu.data = NA,
  studentid = NULL,
  passageid = NULL,
  season = NULL,
  grade = NULL,
  numwords.p = NULL,
  wrc = NULL,
  time = NULL,
  cases = NULL,
  external = NULL,
  parallel = T,
  n.chains = NA,
  iter = NA,
  burn = NA,
  thin = 1
)
}
\arguments{
\item{calib.data}{- mcem class object}

\item{stu.data}{- student reading data}

\item{studentid}{The column name in the data that represents the unique student identifier.}

\item{passageid}{The column name in the data that represents the unique passage identifier.}

\item{season}{The column name in the data that represents the period of data collection.}

\item{grade}{The column name in the data that represents the grade of student.}

\item{numwords.p}{The column name in the data that represents the number of words in a passage.}

\item{wrc}{The column name in the data that represents the words read correctly for each case.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{cases}{- student id vectors, will directly use passage data if no calib.data provided}

\item{external}{- if not NULL, will use not student read passages for estimating}

\item{parallel}{parallel=T, #logical, run in parallel? "T" or "F"}

\item{n.chains}{int., number of the chains}

\item{iter}{int., number of the iterations after the burn-in period}

\item{burn}{int., number of the burn-in iteration}

\item{thin}{int, thinning interval, a.k.a, period of saving samples}
}
\value{
list
}
\description{
Bayes function when running mcem with mcmc setting
}
