% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{exclude_passages}
\alias{exclude_passages}
\title{To exclude error passages}
\usage{
exclude_passages(passage)
}
\arguments{
\item{passage}{}
}
\value{
passage data set without error passages
}
\description{
To exclude error passages
}
