% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get.cases}
\alias{get.cases}
\title{The function get return cases used for wcpm function}
\usage{
get.cases(data)
}
\arguments{
\item{data}{= student response data}
}
\value{
cases vector
}
\description{
The function get return cases used for wcpm function
}
