% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get.perfectcases}
\alias{get.perfectcases}
\title{The function get perfect accurate cases}
\usage{
get.perfectcases(data)
}
\arguments{
\item{data}{= student response data}
}
\value{
perfect accurate case vector
}
\description{
The function get perfect accurate cases
}
