% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getCases}
\alias{getCases}
\title{The function get return cases used for wcpm function}
\usage{
getCases(data)
}
\arguments{
\item{data}{= student response data}
}
\value{
cases vector
}
\description{
The function get return cases used for wcpm function
}
