% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getPerfectcase}
\alias{getPerfectcase}
\title{The function get perfect accurate cases}
\usage{
getPerfectcase(data)
}
\arguments{
\item{data}{= student response data}
}
\value{
perfect accurate case vector
}
\description{
The function get perfect accurate cases
}
