% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{log.initiating}
\alias{log.initiating}
\title{This file includes the log functions of orfr package.}
\usage{
\method{log}{initiating}()
}
\description{
Copyright (C) 2021 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/

This is log.initiating function, which is used to log output.
}
