% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{mcem}
\alias{mcem}
\title{This is an interface function to call and run mcem.}
\usage{
mcem(
  object,
  K.in = 5,
  reps.in = 2,
  ests.in,
  data_check = FALSE,
  est = "MCEM",
  se = "Analytic",
  verbose = FALSE
)
}
\arguments{
\item{object}{- reading data object comes from prep function}

\item{K.in}{- number of passages, default is 5}

\item{reps.in}{- repeats, default is 2}

\item{ests.in}{- if not give, mom function will be called and get est.in output}

\item{data_check}{- boolean, if need to have a data check, default is FALSE}

\item{est}{- estimator keyword, MCEM or MCMC}

\item{se}{- standard error keyword, default is Analytic}

\item{verbose}{- boolean, if shows the summary, default is FALSE}
}
\value{
MCEM list, MCMC list
}
\description{
Copyright (C) 2021 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
