% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{preplong}
\alias{preplong}
\title{The function returns long format data for wcpm function}
\usage{
preplong(data, studentid, passageid, season, grade, nwords.p, wrc, time)
}
\arguments{
\item{data}{= student response data}
}
\value{
data frame
}
\description{
The function returns long format data for wcpm function
}
