% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{prepwide}
\alias{prepwide}
\title{The function returns wide format data for mcpm function}
\usage{
prepwide(data, studentid, passageid, nwords.p, wrc, time)
}
\arguments{
\item{data}{= student response data}
}
\value{
list
}
\description{
The function returns wide format data for mcpm function
}
