% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{run.mcem}
\alias{run.mcem}
\title{This file includes the Base functions of orfr package.}
\usage{
run.mcem(Y, logT10, N, I, k.in = 5, reps.in = 2, ests.in = NA, verbose = FALSE)
}
\arguments{
\item{Y}{= n x I matrix of reading scores -- missingness allowed}

\item{logT10}{= n x I matrix of log10(reading times) -- missingness allowed}

\item{N}{= vector of passage lengths}

\item{I}{= number of passages}

\item{k.in}{= number of imputations, default is 5}

\item{reps.in}{= number of Monte-Carlo iterations, default is 2}

\item{ests.in}{= if not give, mom function will be called and get est.in output}

\item{verbose}{- boolean, if shows the summary, default is FALSE}
}
\value{
mcem list
a,b = parameters controlling binomial success probabilities, each length I
alpha,beta = parameters controlling reading times, each length I
var_tau = variance of latent reading ability tau
rho = correlation between two latent variables
}
\description{
Copyright (C) 2021 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/


This is run.mcem function.
Update Memo:
04/29/2021 Modified the mcem function
           based on Nelis's updated.
10/28/2021 Modified the mcem output
}
