% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{run.scoring}
\alias{run.scoring}
\title{This is run.wcpm function.}
\usage{
run.scoring(
  object,
  person.data,
  task.data,
  cases,
  perfect.cases,
  est = "map",
  lo = -4,
  hi = 4,
  q = 100,
  kappa = 1,
  external = NULL,
  type = NULL
)
}
\arguments{
\item{object}{= fit.model object, if not be given will occur error and stop running}

\item{person.data}{= individual response task data}

\item{task.data}{= estimate parameters data}

\item{cases}{= individual occasion id vector}

\item{perfect.cases}{= perfect accurate case}

\item{est}{= estimator, c("mle", "map", "eap", "mcmc"), default "map"}

\item{lo}{= default -4}

\item{hi}{= default 4}

\item{q}{= default 100}

\item{kappa}{= default 1}

\item{external}{= if not NULL, will use not student read passages for estimating}

\item{type}{- output type, "general" and "orf", default "general" only output tau & theta. "orf" will output wcpm}
}
\value{
wcpm list
}
\description{
Update Memo:
04/29/2021 Modified the wcpm function
           based on Nelis's updated.
06/01/2021 Modified based on Nelis's MAP function.
06/01/2021 Modified based on Sarunya's BiEAP function.
06/20/2021 Modified based on Nelis's updated for MLE and EAP
06/21/2021 Modified a bug of MAP function.
07/12/2021 Modified est.eqs function based on Nelis's code.
07/13/2021 Added Map function for bootstrap.
07/30/2021 Modified wcpm function based on Sarunya's update
}
