% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{scoring}
\alias{scoring}
\title{This is an interface function to call and run wcpm or bootstrap.}
\usage{
scoring(
  calib.data = NA,
  person.data = NA,
  person.id = "",
  task.id = "",
  occasion = "",
  group = "",
  max.counts = "",
  obs.counts = "",
  time = "",
  cases = NULL,
  est = "map",
  se = "analytic",
  failsafe = 0,
  bootstrap = 100,
  external = NULL,
  type = "general"
)
}
\arguments{
\item{calib.data}{- fit.model class object}

\item{person.data}{- individual reading data}

\item{person.id}{The column name in the data that represents the unique individual identifier.}

\item{task.id}{The column name in the data that represents the unique task identifier.}

\item{occasion}{The column name in the data that represents the unique occasion.}

\item{group}{The column name in the data that represents the unique group.}

\item{max.counts}{The column name in the data that represents the number of words in a task.}

\item{obs.counts}{The column name in the data that represents the words read correctly for each case.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{cases}{- individual id vectors, will directly use task data if no calib.data provided}

\item{est}{- estimator keyword / c("mle", "map", "eap", "mcmc"), default is mcmc}

\item{se}{- standard error keyword / c("analytic", "bootstrap"), default is analytic}

\item{failsafe}{- retry time for bootstrap / default 0, can set to 5 ~ 50}

\item{external}{- if not NULL, will use not student read passages for estimating}

\item{type}{- output type, "general" and "orf", default "general" only output tau & theta. "orf" will output wcpm}

\item{bootstrp}{- set K number of bootstrap / default 100}
}
\value{
scoring list or Bootstrap dataset
}
\description{
Copyright (C) 2021 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
