% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bootstrap}
\alias{summary.bootstrap}
\title{summary the information of bootstrap class}
\usage{
summary.bootstrap(
  object,
  digits = 4,
  geterror = FALSE,
  verbose = TRUE,
  factor.scores = FALSE
)
}
\arguments{
\item{object}{= bootstrap object}

\item{digits}{= print out numeric with specific digits}

\item{geterror, }{summary error case, default FALSE}

\item{verbose}{show summary on screen, default TRUE}

\item{factor.scores}{- theta and tau output flag, default is FALSE}
}
\value{
table
}
\description{
Copyright (C) 2021 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
