% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_function.R
\name{bayes}
\alias{bayes}
\title{Bayes function when running mcem with mcmc setting}
\usage{
bayes(
  stu.data = NA,
  studentid = "",
  passageid = "",
  numwords.p = "",
  wrc = "",
  time = "",
  parallel = T,
  n.chains = NA,
  thin = 1,
  iter = NA,
  burn = NA
)
}
\arguments{
\item{stu.data}{- student reading data}

\item{studentid}{The column name in the data that represents the unique student identifier.}

\item{passageid}{The column name in the data that represents the unique passage identifier.}

\item{numwords.p}{The column name in the data that represents the number of words in a passage.}

\item{wrc}{The column name in the data that represents the words read correctly for each case.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{parallel}{parallel=T, #logical, run in parallel? "T" or "F"}

\item{n.chains}{int., number of the chains}

\item{thin}{int, thinning interval, a.k.a, period of saving samples}

\item{iter}{int., number of the iterations after the burn-in period}

\item{burn}{int., number of the burn-in iterations}
}
\value{
list
}
\description{
Bayes function when running mcem with mcmc setting
}
