% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get.cases}
\alias{get.cases}
\title{Returns cases (person and occasion) applied in [fit.model] function.}
\usage{
get.cases(data)
}
\arguments{
\item{data}{= person response data}
}
\value{
cases vector
}
\description{
Returns cases (person and occasion) applied in [fit.model] function.
}
