% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get.cases}
\alias{get.cases}
\title{Returns cases (student and occasion) applied in [wcpm] function.}
\usage{
get.cases(data)
}
\arguments{
\item{data}{= student response data}
}
\value{
cases vector
}
\description{
Returns cases (student and occasion) applied in [wcpm] function.
}
