% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get.perfectcases}
\alias{get.perfectcases}
\title{Returns perfect cases (student and occasion) in which every word was read correctly.}
\usage{
get.perfectcases(data)
}
\arguments{
\item{data}{= student response data}
}
\value{
perfect accurate case vector
}
\description{
Returns perfect cases (student and occasion) in which every word was read correctly.
}
