% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{mcem}
\alias{mcem}
\title{This is an interface function to call and run mcem.}
\usage{
mcem(
  data = NA,
  stu.data = NA,
  studentid = "",
  passageid = "",
  numwords.p = "",
  wrc = "",
  time = "",
  k.in = 5,
  reps.in = 2,
  ests.in = NA,
  est = "mcem",
  se = "analytic",
  verbose = FALSE
)
}
\arguments{
\item{data}{- reading data object comes from prep function}

\item{stu.data}{- student reading data}

\item{studentid}{The column name in the data that represents the unique student identifier.}

\item{passageid}{The column name in the data that represents the unique passage identifier.}

\item{numwords.p}{The column name in the data that represents the number of words in a passage.}

\item{wrc}{The column name in the data that represents the words read correctly for each case.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{k.in}{- number of imputations, default is 5}

\item{reps.in}{- number of Monte-Carlo iterations, default is 2}

\item{ests.in}{- if not given, mom function will be called and get est.in output}

\item{est}{- estimator keyword, mcem or mcmc}

\item{se}{- standard error keyword, analytic or bootstrap, default is analytic,}

\item{verbose}{- boolean, if shows the summary, default is FALSE}
}
\value{
MCEM list, MCMC list
}
\description{
Copyright (C) 2021 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
