% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{preplong}
\alias{preplong}
\title{Prepares data in a long format for [wcpm]}
\usage{
preplong(data, studentid, passageid, season, grade, numwords.p, wrc, time)
}
\arguments{
\item{data}{= student response data}
}
\value{
data frame
}
\description{
Prepares data in a long format for [wcpm]
}
