% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{preplong}
\alias{preplong}
\title{Prepares data in a long format for [wcpm]}
\usage{
preplong(
  data,
  person.id,
  task.id,
  occasion,
  group,
  max.counts,
  obs.counts,
  time
)
}
\arguments{
\item{data}{= person response data}
}
\value{
data frame
}
\description{
Prepares data in a long format for [wcpm]
}
