% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{prepwide}
\alias{prepwide}
\title{Prepares data in a wide format for [mcem].}
\usage{
prepwide(data, person.id, task.id, max.counts, obs.counts, time)
}
\arguments{
\item{data}{A data frame.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{studentid}{The column name in the data that represents the unique student identifier.}

\item{passageid}{The column name in the data that represents the unique passage identifier.}

\item{numwords.p}{The column name in the data that represents the number of words in a passage.}

\item{wrc}{The column name in the data that represents the words read correctly for each case.}
}
\description{
This function will return a list with 5 elements:
Y: a matrix of words read correctly, where rows represent cases (student and occasion) and columns represent passages
logt10: a [tibble::tibble()] of words read correctly
N: the number of cases per passage
I: the number of passages
}
