% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{wcpm}
\alias{wcpm}
\title{This is an interface function to call and run wcpm or bootstrap.}
\usage{
wcpm(
  object,
  stu.data,
  pass.data = NA,
  cases = NA,
  est = "MAP",
  se = "Analytic",
  wo = "internal",
  failsafe = 0,
  bootstrap = 100
)
}
\arguments{
\item{object}{- mcem class object}

\item{stu.data}{- student reading data}

\item{pass.data}{- passage parameters data}

\item{cases}{- student id vectors}

\item{est}{- estimator keyword / c("MLE", "MAP", "EAP")}

\item{se}{- standard error keyword / c("Analytic", "bootstrp")}

\item{wo}{- wcpm option / c("internal", "external"), default is internal}

\item{failsafe}{- retry time for bootstrp / default 0, can set to 5 ~ 50}

\item{bootstrp}{- set K number of bootstrp / default 100}
}
\value{
WCPM list or Bootstrap dataset
}
\description{
Copyright (C) 2021 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
