% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{wcpm}
\alias{wcpm}
\title{This is an interface function to call and run wcpm or bootstrap.}
\usage{
wcpm(
  calib.data = NA,
  stu.data = NA,
  studentid = "",
  passageid = "",
  season = "",
  grade = "",
  numwords.p = "",
  wrc = "",
  time = "",
  cases = NA,
  est = "map",
  se = "analytic",
  wo = "internal",
  failsafe = 0,
  bootstrap = 100,
  external = NULL
)
}
\arguments{
\item{calib.data}{- mcem class object}

\item{stu.data}{- student reading data}

\item{studentid}{The column name in the data that represents the unique student identifier.}

\item{passageid}{The column name in the data that represents the unique passage identifier.}

\item{numwords.p}{The column name in the data that represents the number of words in a passage.}

\item{wrc}{The column name in the data that represents the words read correctly for each case.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{cases}{- student id vectors, will directly use passage data if no calib.data provided}

\item{est}{- estimator keyword / c("mle", "map", "eap")}

\item{se}{- standard error keyword / c("analytic", "bootstrap")}

\item{wo}{- wcpm option / c("internal", "external"), default is internal}

\item{failsafe}{- retry time for bootstrap / default 0, can set to 5 ~ 50}

\item{external}{- if not NULL, will use not student read passages for estimating}

\item{bootstrp}{- set K number of bootstrap / default 100}
}
\value{
WCPM list or Bootstrap dataset
}
\description{
Copyright (C) 2021 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
